fix: add unit + integration tests for payments and social feed
tests/unit/test_payments.py (21 tests):
- Rate feed: TTL cache, CoinGecko/Binance fallback, hardcoded fallback,
  coin_per_usd conversion, unsupported coin error
- Webhook HMAC: valid/invalid signature, dev/prod mode behaviour
- DB helpers: save/load/update roundtrip, rollback on error
- Auto-expiry: expired payments marked on status poll
- Rates endpoint: live rates structure, 503 on feed failure

tests/unit/test_social_feed.py (27 tests):
- _publish_signal(): item creation, reactions/comments init, UUID generation
- GET /api/feed: empty feed, pagination, symbol filter, sort order, private exclusion
- POST /react: thumbs up/down, toggle off, reaction change, 404, 422
- POST /comment: creation, count increment, 404, empty/too-long validation
- GET /comments: empty list, all comments, 404
- POST /opt-in / opt-out: add/remove user, idempotency

tests/integration/test_payments_integration.py (13 tests):
- CryptoPayment: create/retrieve, unique constraint, to_dict(), status update,
  webhook_payload JSON storage
- OutboxEvent: create, published_at update, unpublished query
- ConfigStore: create/retrieve, unique key, value update
- Rate feed: positive prices, USDT ≈ 1.0

Also fix CryptoPayment.id and OutboxEvent.id to use Integer (not BigInteger)
for SQLite compatibility in tests (BigInteger maps to INTEGER in SQLite anyway).

Fixes #12

+++ database/models.py
@@ -825,7 +825,8 @@ if SQLALCHEMY_AVAILABLE:
 
         __tablename__ = "crypto_payments"
 
-        id = Column(BigInteger, primary_key=True, autoincrement=True)
+        # Use Integer for SQLite compatibility (BigInteger maps to INTEGER in SQLite anyway)
+        id = Column(Integer, primary_key=True, autoincrement=True)
         payment_id = Column(String(100), unique=True, nullable=False, index=True)
         user_id = Column(String(128), nullable=False, index=True)
         plan_id = Column(String(100), nullable=False)
@@ -890,7 +891,7 @@ if SQLALCHEMY_AVAILABLE:
 
         __tablename__ = "outbox_events"
 
-        id = Column(BigInteger, primary_key=True, autoincrement=True)
+        id = Column(Integer, primary_key=True, autoincrement=True)
         event_type = Column(String(100), nullable=False, index=True)
         channel = Column(String(100), nullable=False)   # Redis pub/sub channel
         payload = Column(Text, nullable=False)           # JSON
